--- Proyecto_Tienda_Grupo3-jiajun_prueba_rama/comandos/AgregarCompra.java
@@ -1,46 +1,0 @@
-package comandos;
-
-
-
-import entidades.Compra;
-import entidades.Producto;
-import lectorArchivos.CompraData;
-import lectorArchivos.ProductoData;
-
-import java.util.List;
-import java.util.Scanner;
-
-public class AgregarCompra implements Comando{
-    Scanner scanner = new Scanner(System.in);
-    List<Producto> productos = ProductoData.leerTodosProductos();
-    List<Compra> compras = CompraData.leerTodasCompras();
-
-    @Override
-    public void ejecutar() {
-        System.out.println("Nombre del cliente: ");
-        String nombreCliente = scanner.nextLine();
-        System.out.println("Nombre producto: ");
-        String nameProducto = scanner.nextLine();
-        boolean findProducto = false;
-
-        for (Producto producto : productos) {
-            if (nameProducto.equals(producto.getNombre())) {
-                findProducto = true;
-
-                System.out.println("Cantidad del producto: ");
-                int cantidadProducto = scanner.nextInt();
-
-                Compra nuevacompra = new Compra(nombreCliente, nameProducto, cantidadProducto);
-                compras.add(nuevacompra);
-                CompraData.guardarCompra(nuevacompra);
-
-                System.out.println("Se ha añadido la compra!");
-                System.out.println(
-                        nombreCliente + " | " + nameProducto + " | " + cantidadProducto + " | " +"\n");
-            }
-        }
-        if (!findProducto) {
-            System.out.println("\n" + "No existe este producto!" + "\n");
-        }
-    }
-}
