--- Proyecto_Tienda_Grupo3-jiajun_prueba_rama/comandos/BuscarProducto.java
@@ -1,31 +1,0 @@
-package comandos;
-
-import entidades.Producto;
-import lectorArchivos.ProductoData;
-
-
-import java.util.List;
-import java.util.Scanner;
-
-public class BuscarProducto implements Comando{
-
-    Scanner scanner;
-    List<Producto> productos = ProductoData.leerTodosProductos();
-
-    @Override
-    public void ejecutar() {
-        System.out.print("Ingresa el nombre del producto a buscar: ");
-        String nombreProducto = scanner.nextLine();
-        boolean encontrado = false;
-        for (Producto producto : productos) {
-            if (producto.getNombre().equalsIgnoreCase(nombreProducto)) {
-                encontrado = true;
-                System.out.println(producto.toString());
-            }
-        }
-        if (!encontrado) {
-            System.out.println("No hay producto con este nombre");
-        }
-    }
-
-}
