--- Proyecto_Tienda_Grupo3-jiajun_prueba_rama/comandos/Comando.java
@@ -1,5 +1,0 @@
-package comandos;
-
-public interface Comando {
-    void ejecutar();
-}
