--- Proyecto_Tienda_Grupo3-jiajun_prueba_rama/comandos/VerCompras.java
@@ -1,20 +1,0 @@
-package comandos;
-
-
-import entidades.Compra;
-import lectorArchivos.CompraData;
-
-import java.util.List;
-
-
-public class VerCompras implements Comando {
-
-
-    List<Compra> compras = CompraData.leerTodasCompras();
-
-    @Override
-    public void ejecutar() {
-        System.out.println("Lista de comandas.csv:");
-        compras.stream().forEach(producto -> System.out.println(producto.toString()));
-    }
-}
