+++ comandos/Comando.java
@@ -1,0 +1,5 @@
+package comandos;
+
+public interface Comando {
+    void ejecutar();
+}
